Commit 03

+++ Water-Transport-System/src/main/java/Team/Gamma/Water_Transport_System/Controller/UserController.java
@@ -40,7 +40,7 @@ public class UserController {
     }
 
     @GetMapping("/details")
-    public User getUserDetails(@RequestParam("userid") int userid) {
+    public User getUserDetails(@RequestParam("userid") Long userid) {
         // Fetch the current user's details from the UserService
         return userService.getUserDetails(userid);
     }
