+++ Water-Transport-System/src/main/java/Team/Gamma/Water_Transport_System/Repository/UserRepo.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 @EnableJpaRepositories
 @Repository
-public interface UserRepo extends JpaRepository<User,Integer>
+public interface UserRepo extends JpaRepository<User,Long>
 {
     Optional<User> findOneByEmailAndPassword(String email, String password);
     User findByEmail(String email);
-    Optional<User> findByUserid(int userid);
+    Optional<User> findByUserid(Long userid);
 }
